Removing expiration time
