+++ quic/s2n-quic-core/src/datagram/mod.rs
@@ -3,5 +3,7 @@
 
 #[cfg(feature = "alloc")]
 pub mod default;
+pub mod disabled;
 pub mod traits;
+pub use disabled::*;
 pub use traits::*;
